requirements.txt
CacheControl==0.14.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==45.0.6
firebase-admin==6.5.0
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
msgpack==1.0.8
numpy==2.3.2
pandas==2.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rsa==4.9
six==1.17.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
sseclient-py==1.8.0
requests==2.32.3
